refactor(core): move card resolution verification into treeship-core (slice 1)
First slice of the unified-verifier consolidation (the report's #1 P1: one
verifier). The certificate-chain walk that backs `resolve --hub` and
`verify-presentation` lived only in the CLI, so the WASM verifier and the SDKs
could not run it. Move it into core so all surfaces share one code path.

- New treeship_core::verify::resolution with chain_verify_card, verifier_from_trust,
  and ChainVerdict, lifted verbatim from packages/cli/src/commands/resolve.rs.
  Pure and time-injected (caller supplies `now` and trust roots; no I/O).
- verify.rs becomes verify/mod.rs to host the submodule.
- The full chain-walk test suite moves with it and passes byte-for-byte in core
  (4 tests: valid chain, unpinned/wrong-kind ship, expired/not-yet-valid/missing
  window, subject/agent/signer/keyid mismatches) — the proof of zero behavior
  change.
- CLI consumers (resolve, present, history, match, capability) now import from
  treeship_core::verify::resolution; the pub(crate) copies are deleted.

No behavior change. WASM/SDK exposure of verify_resolution is slice 2.
cargo build clean; 147 CLI tests + 59 core verify tests pass; fmt clean.
